Remove CodecProvider dependency on Internal classes. Move AbstractTestStore to better package, to be relocated to versioned/tests once that won't introduce a circular dependency. Some cleanup of the code.
