+++ versioned/mongodb/src/test/java/com/dremio/nessie/versioned/store/mongodb/TestEntityToBsonConverter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.dremio.nessie.versioned.impl;
+package com.dremio.nessie.versioned.store.mongodb;
 
 import java.io.StringWriter;
 import java.util.Map;
@@ -23,6 +23,7 @@ import org.bson.json.JsonWriter;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import com.dremio.nessie.versioned.impl.SampleEntities;
 import com.dremio.nessie.versioned.store.Entity;
 import com.google.common.collect.ImmutableMap;
